- Update some methods in tetris_board
- Use .at() for std::array
- Use .test() for std::bitset

+++ core/tetris_board.hpp
@@ -10,12 +10,12 @@ namespace tetris {
 /**
  * @brief Maximum height of a Tetris board
  */
-constexpr std::int64_t maxHeight{40};
+constexpr std::int32_t maxHeight{40};
 
 /**
  * @brief Maximum width of a Tetris board
  */
-constexpr std::int64_t maxWidth{32};
+constexpr std::int32_t maxWidth{32};
 
 /**
  * @class Board
@@ -32,17 +32,17 @@ public:
    * @param width The width of the board
    * @param height The height of the board
    */
-  Board(std::int64_t width, std::int64_t height);
+  Board(std::int32_t width, std::int32_t height);
 
   /**
    * @brief Equality operator
    */
-  bool operator==(const Board &other) const;
+  bool operator==(const Board& other) const;
 
   /**
    * @brief Inequality operator
    */
-  bool operator!=(const Board &other) const;
+  bool operator!=(const Board& other) const;
 
   /**
    * @brief Fill a cell at the given coordinates
@@ -50,7 +50,7 @@ public:
    * @param x X-coordinate (column)
    * @param y Y-coordinate (row)
    */
-  void fillCell(std::int64_t x, std::int64_t y);
+  void fillCell(std::int32_t x, std::int32_t y);
 
   /**
    * @brief Clear a cell at the given coordinates
@@ -58,35 +58,35 @@ public:
    * @param x X-coordinate (column)
    * @param y Y-coordinate (row)
    */
-  void clearCell(std::int64_t x, std::int64_t y);
+  void clearCell(std::int32_t x, std::int32_t y);
 
   /**
    * @brief Get the width of the board
    *
    * @return The width
    */
-  [[nodiscard]] std::int64_t getWidth() const { return m_width; }
+  [[nodiscard]] std::int32_t getWidth() const { return m_width; }
 
   /**
    * @brief Get the height of the board
    *
    * @return The height
    */
-  [[nodiscard]] std::int64_t getHeight() const { return m_height; }
+  [[nodiscard]] std::int32_t getHeight() const { return m_height; }
 
   /**
    * @brief Get the current highest filled cell in the board
    *
    * @return The height of the highest filled cell
    */
-  [[nodiscard]] std::int64_t getRoof() const { return m_roof; }
+  [[nodiscard]] std::int32_t getRoof() const { return m_roof; }
 
   /**
    * @brief Get the number of filled cells in the board
    *
    * @return The count of filled cells
    */
-  [[nodiscard]] std::int64_t getFilledCellCount() const {
+  [[nodiscard]] std::int32_t getFilledCellCount() const {
     return m_filledCellCount;
   }
 
@@ -96,14 +96,14 @@ public:
    * @param column The column index
    * @return The height of the highest filled cell in the column
    */
-  [[nodiscard]] std::int64_t getColumnHeight(std::int64_t column) const;
+  [[nodiscard]] std::int32_t getColumnHeight(std::int32_t column) const;
 
   /**
    * @brief Clear filled rows and update the board state
    *
    * @return The number of rows cleared
    */
-  std::int64_t clearFilledRows();
+  std::int32_t clearFilledRows();
 
   /**
    * @brief Check if a row is completely filled
@@ -111,21 +111,21 @@ public:
    * @param row The row index
    * @return true if the row is filled, false otherwise
    */
-  [[nodiscard]] bool isRowFilled(std::int64_t row) const;
+  [[nodiscard]] bool isRowFilled(std::int32_t row) const;
 
   /**
    * @brief Get a read-only reference to the board cells
    *
    * @return A reference to the bitset representing the board
    */
-  [[nodiscard]] const std::bitset<maxWidth * maxHeight> &getCells() const;
+  [[nodiscard]] const std::bitset<maxWidth * maxHeight>& getCells() const;
 
   /**
    * @brief Get a read-only view of the column heights
    *
    * @return A span of the column heights
    */
-  [[nodiscard]] std::span<const std::int64_t> getColumnHeights() const;
+  [[nodiscard]] std::span<const std::int32_t> getColumnHeights() const;
 
 private:
   /**
@@ -138,17 +138,17 @@ private:
    *
    * @param column The column to update
    */
-  void updateHeights(std::int64_t column);
+  void updateHeights(std::int32_t column);
 
 private:
   std::bitset<maxWidth * maxHeight>
       m_cells{}; ///< Bit representation of the board
-  std::array<std::int64_t, maxWidth>
+  std::array<std::int32_t, maxWidth>
       m_columnHeights{};            ///< Height of each column
-  std::int64_t m_width{};           ///< Width of the board
-  std::int64_t m_height{};          ///< Height of the board
-  std::int64_t m_roof{};            ///< Current highest filled cell
-  std::int64_t m_filledCellCount{}; ///< Number of filled cells
+  std::int32_t m_width{};           ///< Width of the board
+  std::int32_t m_height{};          ///< Height of the board
+  std::int32_t m_roof{};            ///< Current highest filled cell
+  std::int32_t m_filledCellCount{}; ///< Number of filled cells
 };
 
 } // namespace tetris
